retrace: Add --loopsecs argument.
The traces-db traces are all trimmed to be one giant setup frame and
one frame of rendering.  If you want to do performance work on that,
you really just care about measuring the rendering frame, not the
setup frame.  People have suggested ranking "--loop=" up high enough,
but it turns out that there's not really a useful value of "high
enough" if the setup time might be also changing (such as when
evaluating zink vs native GL).

Sample output:

  eglretrace ~/src/traces-db/valve/portal-2-v2.trace  --loopsecs=5
  warning: --loop blindly repeats the last frame calls, therefore frames might not necessarily render correctly (https://github.com/apitrace/apitrace/issues/800)
[...]
  Rendered the last frame 2305 times in 5.00101 secs, average of 460.907 fps
  Rendered 2308 frames in 7.15039 secs, average of 322.78 fps

XXX: This doesn't glFinish() at the end, so it's relying on
swapbuffers throttling having made the rendering actually complete.